1/2 lay lich

+++ Quoc_Bao_OOP/Standard_room.java
@@ -1,5 +1,5 @@
 public class Standard_room extends Room {
-    public Standard_room(String name, int size, int status,float price,int tang){
+    public Standard_room(String name, int size, int status,double price,int tang){
         super(name, size, status, price,tang);
     }
