+++ Quoc_Bao_OOP/Vip_room.java
@@ -1,5 +1,5 @@
 public class Vip_room extends Room{
-    public Vip_room(String name, int size, int status,float price,int tang){
+    public Vip_room(String name, int size, int status,double  price,int tang){
         super(name, size, status, price,tang);
     }
